+++ Assets/Scenes/BuildScenes/Submenus/CreditsScreen.unity
@@ -12798,6 +12798,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1668760145}
   - component: {fileID: 1668760146}
+  - component: {fileID: 1668760147}
   m_Layer: 5
   m_Name: MovingUI
   m_TagString: Untagged
@@ -12842,8 +12843,35 @@ MonoBehaviour:
   _screenScrollSpeed: 250
   _scrollWaitTime: 2
   _sceneCanvas: {fileID: 814277661}
+--- !u!114 &1668760147
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1668760144}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f939fe2da09f42449ac93937f7da5494, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   _skipPromptTextAnimator: {fileID: 1938957593}
   _skipPromptDuration: 1.5
+  _onSkip:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 1668760146}
+        m_TargetAssemblyTypeName: CreditsScrolling, Assembly-CSharp
+        m_MethodName: ReturnToMainMenu
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
 --- !u!1 &1673243805
 GameObject:
   m_ObjectHideFlags: 0
