+++ Assets/Scripts/UI/CreditsCinematicsSkipUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f939fe2da09f42449ac93937f7da5494
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
